requirements.txt
bittensor==9.7.0
python-dotenv
pandas
plotly
numpy